+++ src/REST/Controller.php
@@ -251,6 +251,13 @@ class Controller extends \WP_REST_Controller
 			'callback'            => array($this, 'post_bmltwf_restore'),
 			'permission_callback' => array($this, 'post_bmltwf_restore_permissions_check'),
 		));
+		
+		// POST options/debug
+		register_rest_route($this->bmltwf_rest_namespace, '/' . $this->bmltwf_options_rest_base . '/debug', array(
+			'methods'             => \WP_REST_Server::CREATABLE,
+			'callback'            => array($this, 'post_bmltwf_debug'),
+			'permission_callback' => array($this, 'post_bmltwf_debug_permissions_check'),
+		));
 	}
 
 	private function authorization_status_code()
@@ -436,6 +443,15 @@ class Controller extends \WP_REST_Controller
 		}
 		return true;
 	}
+	
+	public function post_bmltwf_debug_permissions_check($request)
+	{
+		$this->debug_log("post_bmltwf_debug_permissions_check " . get_current_user_id());
+		if (!current_user_can('manage_options')) {
+			return new \WP_Error('rest_forbidden', __('Access denied: You cannot download debug logs.','bmlt-workflow'), array('status' => $this->authorization_status_code()));
+		}
+		return true;
+	}
 
 	public function post_submissions_permissions_check($request)
 	{
@@ -545,6 +561,12 @@ class Controller extends \WP_REST_Controller
 	public function post_bmltwf_restore($request)
 	{
 		$result = $this->OptionsHandler->post_bmltwf_restore_handler($request);
+		return rest_ensure_response($result);
+	}
+	
+	public function post_bmltwf_debug($request)
+	{
+		$result = $this->OptionsHandler->post_bmltwf_debug_handler($request);
 		return rest_ensure_response($result);
 	}
 }